refactor lofi modules
