+++ src-tauri/python/lofi/__init__.py
@@ -1,0 +1,1 @@
+"""Lofi audio generation utilities."""
