+++ src-tauri/tauri.conf.json
@@ -35,7 +35,7 @@
       "icons/icon.ico"
     ],
     "resources": [
-      "python/lofi_gpu_hq.py"
+      "python/lofi"
     ]
 
   }
